refactor issue filter
